feat: code ui CartHeader and custom hook useSearchProduct
