+++ ShopeeClone/src/Components/CartHeader/index.ts
@@ -1,0 +1,3 @@
+import CartHeader from './CartHeader'
+
+export default CartHeader
