+++ ShopeeClone/src/Components/NavHeader/index.ts
@@ -1,0 +1,3 @@
+import NavHeader from './NavHeader'
+
+export default NavHeader
