+++ ShopeeClone/src/layouts/CartLayout/index.ts
@@ -1,0 +1,3 @@
+import CartHeader from 'src/Components/CartHeader'
+
+export default CartHeader
