+++ ShopeeClone/src/pages/Header/Header.tsx
@@ -1,52 +1,22 @@
-import { yupResolver } from '@hookform/resolvers/yup'
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
-import { omit } from 'lodash'
 import { useContext } from 'react'
-import { useForm } from 'react-hook-form'
-import { createSearchParams, Link, useNavigate } from 'react-router-dom'
-import { toast } from 'react-toastify'
-import authApi from 'src/apis/auth.api'
+import { Link } from 'react-router-dom'
 import purchaseAPI from 'src/apis/purchase.api'
 import Popover from 'src/Components/Popover'
 import path from 'src/constants/path'
 import { PurchaseStatus } from 'src/constants/purchase'
 import { AppContext } from 'src/contexts/app.context'
-import useQueryConfig from 'src/hooks/useQueryConfig'
-import { type Schema, schema } from 'src/utils/rules'
 import emptyCart from 'src/assets/images/empty-cart.webp'
 import { formatCurrency } from 'src/utils/utils'
 import type { Purchase } from 'src/types/purchase.type'
-
-type FormData = Pick<Schema, 'name'>
-const nameSchema = schema.pick(['name'])
+import NavHeader from 'src/Components/NavHeader'
+import { useQuery } from '@tanstack/react-query'
+import useSearchProduct from 'src/hooks/useSearchProduct'
 
 const DISPLAY_MAX = 5
 
 export default function Header() {
-  const queryConfig = useQueryConfig()
-  const queryClient = useQueryClient()
-  const navigate = useNavigate()
-  const { register, handleSubmit } = useForm<FormData>({
-    defaultValues: {
-      name: ''
-    },
-    resolver: yupResolver(nameSchema)
-  })
-  const { setIsAuthenticated, isAuthenticated, setProfile, profile } = useContext(AppContext)
-  const logoutMutation = useMutation({
-    mutationFn: () => authApi.logoutAccount(),
-    onSuccess: () => {
-      toast.success('Đăng xuất thành công !')
-      setIsAuthenticated(false)
-      queryClient.removeQueries({ queryKey: ['purchases', { status: PurchaseStatus.inCart }] })
-    }
-  })
-
-  const handleLogout = () => {
-    logoutMutation.mutate()
-
-    setProfile(null) // Xóa thông tin người dùng khỏi context
-  }
+  const { isAuthenticated } = useContext(AppContext)
+  const { handleSubmitSearch, register } = useSearchProduct()
 
   // useQUery sẽ không gọi lại khi component chỉ re-render, nó chỉ gọi lại khi compnt unmount rồi mount lại hoặc giá trị object trong queryKey thay đổi
   const { data: productInCartData } = useQuery({
@@ -57,131 +27,13 @@ export default function Header() {
 
   const productIncart = productInCartData?.data.data as Purchase[]
 
-  const handleSubmitSearch = handleSubmit((data) => {
-    console.log('data search: ', data)
-    const config = queryConfig.order
-      ? omit(
-          {
-            ...queryConfig,
-            name: data.name
-          },
-          ['order', 'sort_by']
-        )
-      : {
-          ...queryConfig,
-          name: data.name
-        }
-    navigate({
-      pathname: path.home,
-      search: createSearchParams(config).toString()
-    })
-  })
-
   return (
     <div className='bg-[linear-gradient(-180deg,#f53d2d,#f63)] pb-5 pt-2 text-white'>
       <div className='container'>
-        <div className='mx-1 flex justify-between sm:justify-end'>
-          {/* Popover language */}
-          <Popover
-            className='flex cursor-pointer items-center py-1 hover:text-white/70 sm:mx-3'
-            renderPopover={
-              <div className='relative rounded-sm border border-gray-200 bg-white shadow-sm'>
-                <div className='pr-15 flex flex-col px-3 py-2 pl-3 sm:pr-28'>
-                  <button className='px-2 py-1 hover:text-customOrange sm:px-3 sm:py-2'>Tiếng Việt</button>
-                  <button className='mt-1 px-2 py-1 hover:text-customOrange sm:mt-2 sm:px-3 sm:py-2'>EngLish</button>
-                </div>
-              </div>
-            }
-          >
-            <svg
-              xmlns='http://www.w3.org/2000/svg'
-              fill='none'
-              viewBox='0 0 24 24'
-              strokeWidth={1.5}
-              stroke='currentColor'
-              className='h-5 w-5'
-            >
-              <path
-                strokeLinecap='round'
-                strokeLinejoin='round'
-                d='M12 21a9.004 9.004 0 0 0 8.716-6.747M12 21a9.004 9.004 0 0 1-8.716-6.747M12 21c2.485 0 4.5-4.03 4.5-9S14.485 3 12 3m0 18c-2.485 0-4.5-4.03-4.5-9S9.515 3 12 3m0 0a8.997 8.997 0 0 1 7.843 4.582M12 3a8.997 8.997 0 0 0-7.843 4.582m15.686 0A11.953 11.953 0 0 1 12 10.5c-2.998 0-5.74-1.1-7.843-2.918m15.686 0A8.959 8.959 0 0 1 21 12c0 .778-.099 1.533-.284 2.253m0 0A17.919 17.919 0 0 1 12 16.5c-3.162 0-6.133-.815-8.716-2.247m0 0A9.015 9.015 0 0 1 3 12c0-1.605.42-3.113 1.157-4.418'
-              />
-            </svg>
-            <span className='mx-1'>Tiếng Việt</span>
-            <svg
-              xmlns='http://www.w3.org/2000/svg'
-              fill='none'
-              viewBox='0 0 24 24'
-              strokeWidth={1.5}
-              stroke='currentColor'
-              className='size-6'
-            >
-              <path strokeLinecap='round' strokeLinejoin='round' d='m19.5 8.25-7.5 7.5-7.5-7.5' />
-            </svg>
-          </Popover>
-
-          {/* Popover account */}
-          {isAuthenticated && (
-            <Popover
-              className='mx-3 flex cursor-pointer items-center py-1 hover:text-white/70'
-              renderPopover={
-                <div className='relative rounded-sm border border-gray-200 bg-white shadow-sm'>
-                  <Link
-                    className='block w-full bg-white px-2 py-2 text-left hover:bg-slate-100 hover:text-customOrange sm:px-4 sm:py-3'
-                    to={path.profile}
-                  >
-                    <span>Tài khoản của tôi</span>
-                  </Link>
-                  <Link
-                    className='block w-full bg-white px-2 py-2 text-left hover:bg-slate-100 hover:text-customOrange sm:px-4 sm:py-3'
-                    to={path.historyCart}
-                  >
-                    <span>Đơn mua</span>
-                  </Link>
-                  <button
-                    onClick={handleLogout}
-                    className='block w-full bg-white px-2 py-2 text-left hover:bg-slate-100 hover:text-customOrange sm:px-4 sm:py-3'
-                  >
-                    <span>Đăng xuất</span>
-                  </button>
-                </div>
-              }
-            >
-              <div className='mr-2 flex flex-shrink-0 items-center justify-center'>
-                <svg
-                  xmlns='http://www.w3.org/2000/svg'
-                  fill='none'
-                  viewBox='0 0 24 24'
-                  strokeWidth={1.5}
-                  stroke='currentColor'
-                  className='mr-1 h-5 w-6 rounded-full object-cover'
-                >
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    d='M15.75 6a3.75 3.75 0 1 1-7.5 0 3.75 3.75 0 0 1 7.5 0ZM4.501 20.118a7.5 7.5 0 0 1 14.998 0A17.933 17.933 0 0 1 12 21.75c-2.676 0-5.216-.584-7.499-1.632Z'
-                  />
-                </svg>
-                <div>{profile ? profile.email || profile.name : 'User'}</div>
-              </div>
-            </Popover>
-          )}
-          {/* before login */}
-          {!isAuthenticated && (
-            <div className='flex items-center'>
-              <Link to={path.register} className='mx-3 capitalize hover:text-white/70'>
-                Đăng Ký
-              </Link>
-              <div className='h-4 border-r-2 border-r-white'></div>
-              <Link to={path.login} className='mx-3 capitalize hover:text-white/70'>
-                Đăng Nhập
-              </Link>
-            </div>
-          )}
-        </div>
+        <NavHeader />
         <div className='mt-4 grid grid-cols-12 items-end gap-4'>
-          <Link to='/' className='col-span-2 ml-2 flex items-center justify-center sm:col-span-2'>
-            <svg viewBox='0 0 192 65' className='h-14 w-14'>
+          <Link to={path.home} className='col-span-2 ml-2 flex items-center justify-center sm:col-span-2'>
+            <svg viewBox='0 0 192 65' className='h-10 w-auto'>
               <g fillRule='evenodd'>
                 <path
                   fill='white'
@@ -196,7 +48,7 @@ export default function Header() {
               <input
                 placeholder='FreeShip Đơn từ 0Đ'
                 type='text'
-                className='h-6 max-w-[170px] flex-grow border-none bg-green-300 px-3 py-2 text-black outline-none placeholder:text-xs sm:h-auto sm:max-w-none sm:placeholder:text-lg'
+                className='h-6 max-w-[170px] flex-grow border-none px-3 py-2 text-black outline-none placeholder:text-xs sm:h-auto sm:max-w-none sm:placeholder:text-lg'
                 {...register('name')}
               />
               <button className='absolute right-1 flex h-6 items-center rounded-sm bg-customOrange px-6 py-2 hover:opacity-90 sm:h-auto'>
